AiSummarizer/requirements.txt
fastapi>=0.115,<1.0
python-dotenv>=1.0,<2.0
uvicorn[standard]>=0.30,<1.0
httpx>=0.27,<1.0
pypdf>=5.0,<6.0
python-docx>=1.1,<2.0
python-multipart>=0.0.9,<1.0
pytest>=8.0,<9.0

